Installing security checking on every request. This is important and making it as easy to code in the future

+++ src/main/java/com/questnr/controllers/CommentActionController.java
@@ -4,11 +4,11 @@ import com.questnr.model.entities.CommentAction;
 import com.questnr.model.projections.CommentActionProjection;
 import com.questnr.requests.CommentActionRequest;
 import com.questnr.services.CommentActionService;
+import com.questnr.services.community.AccessService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -18,6 +18,9 @@ public class CommentActionController {
     @Autowired
     CommentActionService commentActionService;
 
+    @Autowired
+    AccessService accessService;
+
     @RequestMapping(value = "/posts/{postId}/comment", method = RequestMethod.GET)
     Page<CommentActionProjection> getAllCommentsByPostId(@PathVariable Long postId, Pageable pageable) {
         return commentActionService.getAllCommentActionByPostId(postId, pageable);
@@ -25,13 +28,24 @@ public class CommentActionController {
 
     @RequestMapping(value = "/posts/{postId}/comment", method = RequestMethod.POST)
     CommentAction createComment(@PathVariable Long postId, @RequestBody CommentActionRequest commentActionRequest) {
-        commentActionRequest.setPostId(postId);
-        return commentActionService.createCommentAction(commentActionRequest);
+        /*
+         * Post Comment Security Checking
+         * */
+        if (accessService.hasAccessToPost(postId)) {
+            commentActionRequest.setPostId(postId);
+            return commentActionService.createCommentAction(commentActionRequest);
+        }
+        return null;
     }
 
     @RequestMapping(value = "/posts/{postId}/comment/{commentId}", method = RequestMethod.DELETE)
     @ResponseStatus(HttpStatus.OK)
     public void deleteComment(@PathVariable Long postId, @PathVariable Long commentId) {
-        commentActionService.deleteCommentAction(postId, commentId);
+        /*
+         * Post Comment Security Checking
+         * */
+        if (accessService.hasAccessToPost(postId)) {
+            commentActionService.deleteCommentAction(postId, commentId);
+        }
     }
 }
